fix: address review comment 4684308134 - split hosted validation tests
