+++ tests/unit/runtime/test_hosted_validation_delegate_parts/test_hosted_validation_delegate_part_002.py
@@ -1,0 +1,358 @@
+"""Hosted validation delegate profile sanitization tests."""
+
+from __future__ import annotations
+
+import json
+from pathlib import Path
+from typing import Any
+
+import httpx
+import pytest
+
+from awf.profiles.models import WorkspaceProfile
+from awf.runtime.hosted_delegation import HostedValidationDelegate
+from tests.unit.runtime.test_hosted_validation_delegate import (
+    _config,
+    _profile_with_runtime_secret,
+    _profile_with_secret_ref,
+    _profile_with_service_secret,
+    _profile_with_service_without_environment,
+)
+
+
+@pytest.mark.unit
+async def test_hosted_validation_strips_credentials_from_pr_identity_urls(
+    tmp_path: Path,
+) -> None:
+    seen: dict[str, Any] = {}
+
+    async def _handler(request: httpx.Request) -> httpx.Response:
+        if request.method == "POST" and request.url.path == "/v1/validation-runs":
+            seen["body"] = json.loads(request.content)
+            return httpx.Response(
+                202,
+                json={
+                    "operation_id": "val_1",
+                    "workspace_id": "ws_hosted",
+                    "operation_url": "/v1/operations/val_1",
+                },
+            )
+        if request.method == "GET" and request.url.path == "/v1/operations/val_1":
+            return httpx.Response(
+                200,
+                json={
+                    "operation_id": "val_1",
+                    "workspace_id": "ws_hosted",
+                    "state": "succeeded",
+                    "commands": [
+                        {
+                            "command": "uv run pytest tests/unit/foo -q",
+                            "returncode": 0,
+                            "duration_seconds": 1.25,
+                            "stdout": "passed\n",
+                            "stderr": "",
+                            "phase": "validate",
+                            "reason_code": "COMMAND_FAILED",
+                        }
+                    ],
+                },
+            )
+        raise AssertionError(f"unexpected request {request.method} {request.url}")
+
+    async with httpx.AsyncClient(transport=httpx.MockTransport(_handler)) as client:
+        delegate = HostedValidationDelegate(
+            _config(),
+            artifacts_dir=tmp_path,
+            client=client,
+        )
+        await delegate.run_profile_phases(
+            workspace_id="ws_hosted",
+            compose_project="unused",
+            compose_file=tmp_path / "missing-compose.yml",
+            profile=WorkspaceProfile(name="hosted-test"),
+            phase_names=("validate",),
+            pr_identity={
+                "repo_url": "https://base-secret-123@github.com/base/repo.git",
+                "head_repo_url": "ssh://git:fork-secret-456@github.com/fork/repo.git",
+                "pr_number": 277,
+            },
+        )
+
+    pr_identity = seen["body"]["pr_identity"]
+    assert pr_identity["repo_url"] == "https://github.com/base/repo.git"
+    assert pr_identity["head_repo_url"] == "ssh://git@github.com/fork/repo.git"
+    body_blob = json.dumps(seen["body"], sort_keys=True)
+    assert "base-secret-123" not in body_blob
+    assert "fork-secret-456" not in body_blob
+
+
+@pytest.mark.unit
+async def test_hosted_validation_preserves_passwordless_ssh_pr_identity_userinfo(
+    tmp_path: Path,
+) -> None:
+    seen: dict[str, Any] = {}
+
+    async def _handler(request: httpx.Request) -> httpx.Response:
+        if request.method == "POST" and request.url.path == "/v1/validation-runs":
+            seen["body"] = json.loads(request.content)
+            return httpx.Response(
+                202,
+                json={
+                    "operation_id": "val_1",
+                    "workspace_id": "ws_hosted",
+                    "operation_url": "/v1/operations/val_1",
+                },
+            )
+        if request.method == "GET" and request.url.path == "/v1/operations/val_1":
+            return httpx.Response(
+                200,
+                json={
+                    "operation_id": "val_1",
+                    "workspace_id": "ws_hosted",
+                    "state": "succeeded",
+                    "commands": [
+                        {
+                            "command": "uv run pytest tests/unit/foo -q",
+                            "returncode": 0,
+                            "duration_seconds": 1.25,
+                            "stdout": "passed\n",
+                            "stderr": "",
+                            "phase": "validate",
+                            "reason_code": "COMMAND_FAILED",
+                        }
+                    ],
+                },
+            )
+        raise AssertionError(f"unexpected request {request.method} {request.url}")
+
+    async with httpx.AsyncClient(transport=httpx.MockTransport(_handler)) as client:
+        delegate = HostedValidationDelegate(
+            _config(),
+            artifacts_dir=tmp_path,
+            client=client,
+        )
+        await delegate.run_profile_phases(
+            workspace_id="ws_hosted",
+            compose_project="unused",
+            compose_file=tmp_path / "missing-compose.yml",
+            profile=WorkspaceProfile(name="hosted-test"),
+            phase_names=("validate",),
+            pr_identity={
+                "repo_url": "ssh://git@github.com/base/repo.git",
+                "head_repo_url": "ssh://git@github.com/fork/repo.git",
+                "pr_number": 277,
+            },
+        )
+
+    pr_identity = seen["body"]["pr_identity"]
+    assert pr_identity["repo_url"] == "ssh://git@github.com/base/repo.git"
+    assert pr_identity["head_repo_url"] == "ssh://git@github.com/fork/repo.git"
+
+
+@pytest.mark.unit
+async def test_hosted_validation_profile_without_service_environment_passes_through(
+    tmp_path: Path,
+) -> None:
+    seen: dict[str, Any] = {}
+
+    async def _handler(request: httpx.Request) -> httpx.Response:
+        if request.method == "POST" and request.url.path == "/v1/validation-runs":
+            seen["body"] = json.loads(request.content)
+            return httpx.Response(
+                202,
+                json={
+                    "operation_id": "val_1",
+                    "workspace_id": "ws_hosted",
+                    "operation_url": "/v1/operations/val_1",
+                },
+            )
+        if request.method == "GET" and request.url.path == "/v1/operations/val_1":
+            return httpx.Response(
+                200,
+                json={
+                    "operation_id": "val_1",
+                    "workspace_id": "ws_hosted",
+                    "state": "succeeded",
+                    "commands": [],
+                },
+            )
+        raise AssertionError(f"unexpected request {request.method} {request.url}")
+
+    async with httpx.AsyncClient(transport=httpx.MockTransport(_handler)) as client:
+        delegate = HostedValidationDelegate(
+            _config(),
+            artifacts_dir=tmp_path,
+            client=client,
+        )
+        await delegate.run_profile_phases(
+            workspace_id="ws_hosted",
+            compose_project="unused",
+            compose_file=tmp_path / "missing-compose.yml",
+            profile=_profile_with_service_without_environment(),
+            phase_names=("validate",),
+        )
+
+    service = seen["body"]["profile"]["services"][0]
+    assert service["name"] == "redis"
+    assert service["image"] == "redis:7"
+    assert service["environment"] == {}
+
+
+@pytest.mark.unit
+async def test_hosted_validation_sanitizes_literal_runtime_environment_secrets(
+    tmp_path: Path,
+) -> None:
+    seen: dict[str, Any] = {}
+
+    async def _handler(request: httpx.Request) -> httpx.Response:
+        if request.method == "POST" and request.url.path == "/v1/validation-runs":
+            seen["body"] = json.loads(request.content)
+            return httpx.Response(
+                202,
+                json={
+                    "operation_id": "val_1",
+                    "workspace_id": "ws_hosted",
+                    "operation_url": "/v1/operations/val_1",
+                },
+            )
+        if request.method == "GET" and request.url.path == "/v1/operations/val_1":
+            return httpx.Response(
+                200,
+                json={
+                    "operation_id": "val_1",
+                    "workspace_id": "ws_hosted",
+                    "state": "succeeded",
+                    "commands": [],
+                },
+            )
+        raise AssertionError(f"unexpected request {request.method} {request.url}")
+
+    async with httpx.AsyncClient(transport=httpx.MockTransport(_handler)) as client:
+        delegate = HostedValidationDelegate(
+            _config(),
+            artifacts_dir=tmp_path,
+            client=client,
+        )
+        await delegate.run_profile_phases(
+            workspace_id="ws_hosted",
+            compose_project="unused",
+            compose_file=tmp_path / "missing-compose.yml",
+            profile=_profile_with_runtime_secret(),
+            phase_names=("validate",),
+        )
+
+    body_blob = json.dumps(seen["body"], sort_keys=True)
+    assert "npm-profile-secret" not in body_blob
+    assert "pkg-token" not in body_blob
+    assert seen["body"]["profile"]["runtime"]["environment"] == {
+        "NPM_TOKEN": "${NPM_TOKEN}",
+        "OLLAMA_HOST": "http://ollama.profile:11434",
+        "PIP_INDEX_URL": "${PIP_INDEX_URL}",
+    }
+
+
+@pytest.mark.unit
+async def test_hosted_validation_sanitizes_literal_service_environment_secrets(
+    tmp_path: Path,
+) -> None:
+    seen: dict[str, Any] = {}
+
+    async def _handler(request: httpx.Request) -> httpx.Response:
+        if request.method == "POST" and request.url.path == "/v1/validation-runs":
+            seen["body"] = json.loads(request.content)
+            return httpx.Response(
+                202,
+                json={
+                    "operation_id": "val_1",
+                    "workspace_id": "ws_hosted",
+                    "operation_url": "/v1/operations/val_1",
+                },
+            )
+        if request.method == "GET" and request.url.path == "/v1/operations/val_1":
+            return httpx.Response(
+                200,
+                json={
+                    "operation_id": "val_1",
+                    "workspace_id": "ws_hosted",
+                    "state": "succeeded",
+                    "commands": [],
+                },
+            )
+        raise AssertionError(f"unexpected request {request.method} {request.url}")
+
+    async with httpx.AsyncClient(transport=httpx.MockTransport(_handler)) as client:
+        delegate = HostedValidationDelegate(
+            _config(),
+            artifacts_dir=tmp_path,
+            client=client,
+        )
+        await delegate.run_profile_phases(
+            workspace_id="ws_hosted",
+            compose_project="unused",
+            compose_file=tmp_path / "missing-compose.yml",
+            profile=_profile_with_service_secret(),
+            phase_names=("validate",),
+        )
+
+    body_blob = json.dumps(seen["body"], sort_keys=True)
+    assert "literal-service-secret" not in body_blob
+    assert seen["body"]["profile"]["services"][0]["environment"] == {
+        "POSTGRES_PASSWORD": "${POSTGRES_PASSWORD}",
+        "POSTGRES_USER": "awf",
+        "EXTERNAL_API_KEY": "${SERVICE_API_KEY}",
+    }
+
+
+@pytest.mark.unit
+async def test_hosted_validation_strips_profile_secret_refs(tmp_path: Path) -> None:
+    seen: dict[str, Any] = {}
+
+    async def _handler(request: httpx.Request) -> httpx.Response:
+        if request.method == "POST" and request.url.path == "/v1/validation-runs":
+            seen["body"] = json.loads(request.content)
+            return httpx.Response(
+                202,
+                json={
+                    "operation_id": "val_1",
+                    "workspace_id": "ws_hosted",
+                    "operation_url": "/v1/operations/val_1",
+                },
+            )
+        if request.method == "GET" and request.url.path == "/v1/operations/val_1":
+            return httpx.Response(
+                200,
+                json={
+                    "operation_id": "val_1",
+                    "workspace_id": "ws_hosted",
+                    "state": "succeeded",
+                    "commands": [],
+                },
+            )
+        raise AssertionError(f"unexpected request {request.method} {request.url}")
+
+    async with httpx.AsyncClient(transport=httpx.MockTransport(_handler)) as client:
+        delegate = HostedValidationDelegate(
+            _config(),
+            artifacts_dir=tmp_path,
+            client=client,
+        )
+        await delegate.run_profile_phases(
+            workspace_id="ws_hosted",
+            compose_project="unused",
+            compose_file=tmp_path / "missing-compose.yml",
+            profile=_profile_with_secret_ref(),
+            phase_names=("validate",),
+        )
+
+    body_blob = json.dumps(seen["body"], sort_keys=True)
+    assert "local-file:///home/user/.awf/secrets/codex.default" not in body_blob
+    assert seen["body"]["profile"]["secrets"] == [
+        {
+            "name": "codex-default",
+            "target": "/run/awf/secrets/codex-default",
+            "kind": "mount",
+            "mode": "ro",
+            "required": True,
+            "provider": "local-file",
+        }
+    ]
